+++ Engine/UiLoader.cpp
@@ -1,6 +1,6 @@
 /*
  * mTroll MIDI Controller
- * Copyright (C) 2007-2008,2010,2015,2018,2020 Sean Echevarria
+ * Copyright (C) 2007-2008,2010,2015,2018,2020,2025 Sean Echevarria
  *
  * This file is part of mTroll.
  *
@@ -483,7 +483,7 @@ UiLoader::LoadOtherStuffAndFinalize(TiXmlElement * pElem)
 		pElem->QueryIntAttribute("ledBrightness", &ledBrightness);
 
 		if (-1 != ledBrightness)
-			mUi->SetHardwareLedIntensity(ledBrightness);
+			mUi->SetHardwareLedIntensity((short)ledBrightness);
 	}
 }
